+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -5,8 +5,8 @@
 ### Features Added
 
 - Added `validate_authority` support for msal client  #22625
-- Added `resource_id` support for user-assigned managed identity
-- Added `ClientAssertionCredential` support
+  - Added `resource_id` support for user-assigned managed identity  #22329
+- Added `ClientAssertionCredential` support  #22887
 
 ### Breaking Changes